Moved addGroupToStructure methods to StructureTools
Also some doc improvements
