+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/utils/SymmetryTools.java
@@ -536,7 +536,7 @@ public class SymmetryTools {
 			Chain prevChain = null;
 			for(int k=res1;k<=res2; k++) {
 				Group g = atoms[k].getGroup();
-				prevChain = addGroupToStructure(s, g, prevChain,true);
+				prevChain = StructureTools.addGroupToStructure(s, g, prevChain,true);
 				repeat.addAll(g.getAtoms());
 			}
 
@@ -548,60 +548,12 @@ public class SymmetryTools {
 			
 			logger.warn("Adding {} ligands to {}",ligands.size(), symmetry.getMultipleAlignment().getStructureIdentifier(i));
 			for( Group ligand : ligands) {
-				prevChain = addGroupToStructure(s, ligand, prevChain,true);
+				prevChain = StructureTools.addGroupToStructure(s, ligand, prevChain,true);
 			}
 
 			repeats.add(s);
 		}
 		return repeats;
-	}
-
-	/**
-	 * Adds a particular group to a structure. A new chain will be created if necessary.
-	 * 
-	 * <p>When adding multiple groups, pass the return value of one call as the
-	 * chainGuess parameter of the next call for efficiency.
-	 * <pre>
-	 * Chain guess = null;
-	 * for(Group g : groups) {
-	 *     guess = addGroupToStructure(s, g, guess );
-	 * }
-	 * </pre>
-	 * @param s structure to receive the group
-	 * @param g group to add
-	 * @param chainGuess (optional) If not null, should be a chain from s. Used
-	 *  to improve performance when adding many groups from the same chain
-	 * @param clone Indicates whether the input group should be cloned before
-	 *  being added to the new chain
-	 * @return the chain g was added to
-	 */
-	public static Chain addGroupToStructure(Structure s, Group g, Chain chainGuess, boolean clone ) {
-		// Find or create the chain
-		String chainId = g.getChainId();
-		assert !chainId.isEmpty();
-		Chain chain;
-		if(chainGuess != null && chainGuess.getId() == chainId) {
-			// previously guessed chain
-			chain = chainGuess;
-		} else {
-			// Try to guess
-			chain = s.getChain(chainId);
-			if(chain == null) {
-				// no chain found
-				chain = new ChainImpl();
-				chain.setId(chainId);
-				chain.setName(g.getChain().getName());
-				s.addChain(chain);
-			}
-		}
-		
-		// Add cloned group
-		if(clone) {
-			g = (Group)g.clone();
-		}
-		chain.addGroup(g);
-		
-		return chain;
 	}
 
 	/**
